+++ tests/test_tga_image.py
@@ -61,46 +61,35 @@ def _make_tga_file(path, img_type, width=2, height=2, bpp=24):
         f.write(header + pixel_data)
 
 
-class TestUtilTga(PublishApiTestBase):
-    """Tests for util_tga module functions."""
+class TestTgaImage(PublishApiTestBase):
+    """Tests for tga_image module functions."""
 
     def test_is_tga_rle(self):
         """
         Ensures is_tga_rle returns True for RLE type 10 and False for
         uncompressed type 2.
         """
-        util_tga = self.app.import_module("tk_multi_publish2").util_tga
+        tga_image = self.app.import_module("tk_multi_publish2").utils.tga_image
         with tempfile.NamedTemporaryFile(
             suffix=".tga"
         ) as rle_path, tempfile.NamedTemporaryFile(suffix=".tga") as non_rle_path:
             _make_tga_file(rle_path.name, img_type=_TGA_IMG_TYPE_RLE)
             _make_tga_file(non_rle_path.name, img_type=_TGA_IMG_TYPE_UNCOMPRESSED)
-            self.assertTrue(util_tga.is_tga_rle(rle_path.name))
-            self.assertFalse(util_tga.is_tga_rle(non_rle_path.name))
+            self.assertTrue(tga_image.is_tga_rle(rle_path.name))
+            self.assertFalse(tga_image.is_tga_rle(non_rle_path.name))
 
-    def test_decode_tga_to_raw_rle(self):
+    def test_tga_to_pixmap_rle(self):
         """
-        Ensures decode_tga_to_raw correctly decodes an RLE TGA file:
+        Ensures tga_to_pixmap correctly decodes an RLE TGA file:
         pixel buffer has the right length and pixel values match the encoded data.
         """
-        util_tga = self.app.import_module("tk_multi_publish2").util_tga
+        tga_image = self.app.import_module("tk_multi_publish2").utils.tga_image
         with tempfile.NamedTemporaryFile(suffix=".tga") as path:
             _make_tga_file(
                 path.name, img_type=_TGA_IMG_TYPE_RLE, width=2, height=2, bpp=24
             )
-            result = util_tga.decode_tga_to_raw(path.name)
-            self.assertEqual(result["width"], 2)
-            self.assertEqual(result["height"], 2)
-            self.assertEqual(result["psize"], 3)
-            self.assertEqual(
-                len(result["pixels"]),
-                result["width"] * result["height"] * result["psize"],
-            )
-            # Each pixel was encoded as _TEST_PIXEL_VALUE repeated for each channel (BGR)
-            self.assertEqual(
-                result["pixels"],
-                _TEST_PIXEL_VALUE
-                * result["width"]
-                * result["height"]
-                * result["psize"],
-            )
+            pixmap = tga_image.tga_to_pixmap(path.name)
+            self.assertIsNotNone(pixmap)
+            self.assertFalse(pixmap.isNull())
+            self.assertEqual(pixmap.width(), 2)
+            self.assertEqual(pixmap.height(), 2)
